Make digests more reliable
Instead of checking 1000 messages max once a day, we now grab all
digests once a day and then send one digest after sending due messages.
And if the digest has a lot of mails (at least 3 people are subscribed
to build_success PER_WEEK, which are a LOT of emails), we send several
mails - 300 per digest.

This makes sure we don't forget digests, the current downside is that
we send digests of all kind if hermesworker restarts. If people complain
(if we happen to restart several times a day PER_WEEK becomes useless)
we need to either have a timestamp in file system or database. But that
is not as simple either as just because we grabed digests doesn't mean
the digests are sent out. One possible measure is to save per (digest)
subscribtion when the last mail was sent out and use that to limit how
often we regrab it.